Fixed a bug and added some plumbing for ordering parts
In the Tests an additional argument was added so that everything runs smoothly. Added a Model that will be used when the orders are create (the model is not yet completed). An incomplete View for Ordering a Part was introduced.
A Validation Attribute was added that the amount of an ordered part is positive.