Some UI changes and availability of parts introduced
A calculation that decides whether a Repair Task may be assigned. It can be only if there are enough parts of each type available. In addition, if a Repair Task is assigned before the current admin assigns it or parts become unavailable an error is shown in the index admin page.